% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_graphing_functions.R
\name{custom_scatter_plot}
\alias{custom_scatter_plot}
\title{Custom Scatter Plot}
\usage{
custom_scatter_plot(data, x, y)
}
\arguments{
\item{data}{The data frame containing the data}

\item{x}{The variable for the x-axis}

\item{y}{The variable for the y-axis}
}
\value{
A ggplot scatter plot
}
\description{
This function creates a scatter plot with custom color and title.
}
